requirements.txt
# Core utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0
arxiv==2.1.0
feedparser==6.0.10

# NLP
nltk==3.9.1
scikit-learn==1.5.2

# Optional ML embeddings (you can remove if not needed yet)
sentence-transformers==3.0.1

# Dashboard + viz
streamlit==1.38.0
matplotlib==3.9.2
plotly==5.24.1
